--- client/v2/export_test.go
@@ -1,8 +1,0 @@
-package client
-
-type (
-	RPCAverage  = rpcAverage
-	FailureRate = failureRate
-)
-
-const EMAAlpha = emaAlpha
